clients: Move issue embeds to issue page
