--- clients/apps/web/src/app/(main)/dashboard/[organization]/(header)/promote/page.tsx
@@ -1,24 +1,0 @@
-import { getServerSideAPI } from '@/utils/api/serverside'
-import { getOrganizationBySlugOrNotFound } from '@/utils/organization'
-import { Metadata } from 'next'
-import ClientPage from './ClientPage'
-
-export async function generateMetadata(): Promise<Metadata> {
-  return {
-    title: 'Promote', // " | Polar is added by the template"
-  }
-}
-
-export default async function Page({
-  params,
-}: {
-  params: { organization: string }
-}) {
-  const api = getServerSideAPI()
-  const organization = await getOrganizationBySlugOrNotFound(
-    api,
-    params.organization,
-  )
-
-  return <ClientPage organization={organization} />
-}
